app/requirements/prod.txt
# Everything needed in production

# Flask
Flask==2.1.2
Werkzeug==2.1.2
click>=7.0

# Database
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.39
psycopg2-binary==2.9.3

# Migrations
Flask-Migrate==3.1.0

# Forms
Flask-WTF==1.0.1
WTForms==3.0.1

# Admin
Flask-Admin==1.6.0

# Deployment
gevent==21.12.0
gunicorn>=19.9.0
supervisor==4.2.4

# Flask Static Digest
Flask-Static-Digest==0.2.1

# Caching
Flask-Caching>=1.7.2

# Debug toolbar
Flask-DebugToolbar==0.13.1

# Environment variable parsing
environs==9.5.0

# API stuff
flask-restx==0.5.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1

# Mail
Flask-Mail==0.9.1

# Parsing stuff
requests==2.28.1
types-requests==2.28.9
beautifulsoup4==4.11.1



